+++ R/quantile_central_limit_theorem.R
@@ -1,19 +1,21 @@
-# =================================== clt =====================================
+# =================================== cltq ====================================
 
-#' Central Limit Theorem (CLT)
+#' Central Limit Theorem (CLT) for sample quantiles
 #'
-#' A movie to illustrate the ideas of the sampling distribution of a mean
-#' and the central limit theorem.
+#' A movie to illustrate the ideas of the sampling distribution of the
+#' sample 100\eqn{p}\% quantile and the central limit theorem for sample
+#' quantiles.
 #'
 #' @param n An integer scalar.  The size of the samples drawn from the
 #'   distribution chosen using \code{distn}.
-#' @param distn A character scalar specifying the distribution from which
-#'   observations are sampled.   Distributions \code{"beta"},
-#'   \code{"binomial"}, \code{"chisq"}, \code{"chi-squared"},
-#'   \code{"exponential"}, \code{"f"}, \code{"gamma"}, \code{"geometric"},
-#'   \code{"gev"}, \code{"gp"}, \code{"hypergeometric"}, \code{"lognormal"},
-#'   \code{"log-normal"}, \code{"negative binomial"}, \code{"normal"},
-#'   \code{"poisson"}, \code{"t"}, \code{"uniform"} and \code{"weibull"} are
+#' @param p A numeric scalar in (0, 1).  The value of \eqn{p}.
+#' @param distn A character scalar specifying the (continuous) distribution
+#'   from which observations are sampled.   Distributions \code{"beta"},
+#'   \code{"chisq"}, \code{"chi-squared"},
+#'   \code{"exponential"}, \code{"f"}, \code{"gamma"},
+#'   \code{"gev"}, \code{"gp"}, \code{"lognormal"},
+#'   \code{"log-normal"}, \code{"normal"},
+#'   \code{"t"}, \code{"uniform"} and \code{"weibull"} are
 #'   recognised, case being ignored.
 #'
 #'   If \code{distn} is not supplied then \code{distn = "exponential"}
@@ -29,9 +31,6 @@
 #'   If \code{distn = "gamma"} then the \code{(shape, rate)}
 #'   parameterisation is used.  If \code{scale} is supplied via \code{params}
 #'   then \code{rate} is inferred from this.
-#'   If \code{distn = "negative binomial"} then the \code{(size, prob)}
-#'   parameterisation is used.  If \code{mu} is supplied via \code{params}
-#'   then \code{prob} is inferred from this (and \code{size}).
 #'   If \code{distn = "beta"} then \code{ncp} is forced to be zero.
 #' @param params A named list of additional arguments to be passed to the
 #'   density function associated with distribution \code{distn}.
@@ -48,9 +47,11 @@
 #'   \code{"gev"} (\code{shape = 0.2}).
 #'   \code{"gamma"} (\code{shape = 2},
 #'   \code{"gp"} (\code{shape = 0.1}),
-#'   \code{"poisson"} (\code{lambda = 5}) and
 #'   \code{"t"} (\code{df = 4}) and
 #'   \code{"weibull"} (\code{shape = 2}).
+#' @param type An integer between 1 and 9. The value of the argument
+#'   \code{type} to be passed to \code{\link[stats]{quantile}} to when
+#'   calculating a sample quantile.
 #' @param panel_plot A logical parameter that determines whether the plot
 #'   is placed inside the panel (\code{TRUE}) or in the standard graphics
 #'   window (\code{FALSE}).  If the plot is to be placed inside the panel
@@ -64,7 +65,7 @@
 #'   (or decreased) after one click of the + (or -) button in the parameter
 #'   window.
 #' @param arrow A logical scalar.  Should an arrow be included to show the
-#'   simulated sample mean from the top plot being placed into the
+#'   simulated sample quantile from the top plot being placed into the
 #'   bottom plot?
 #' @param leg_cex The argument \code{cex} to \code{\link[graphics]{legend}}.
 #'   Allows the size of the legend to be controlled manually.
@@ -73,66 +74,64 @@
 #'   \code{\link[rpanel]{rp.doublebutton}}, not including \code{panel},
 #'   \code{variable}, \code{title}, \code{step}, \code{action}, \code{initval},
 #'   \code{range}.
-#' @details Loosely speaking, a consequence of the
-#'   \href{https://en.wikipedia.org/wiki/Central_limit_theorem}{Central Limit Theorem}
-#'   is that the mean of a \strong{large number} of independent and
-#'   identically distributed random variables, each with mean \eqn{\mu} and
-#'   finite standard deviation \eqn{\sigma}, has \strong{approximately} a
-#'   normal distribution, even if these original variables are not normally
-#'   distributed.
+#' @details Loosely speaking, a consequence of the CLT for sample quantiles
+#'   is that the 100\eqn{p}\% sample quantile of a \strong{large number} of
+#'   identically distributed random variables, each with probability density
+#'   function \eqn{f} and 100\eqn{p}\% quantile \eqn{\xi(p)}, has
+#'   \strong{approximately} a normal distribution.  See, for example,
+#'   \href{http://dx.doi.org/10.1007/b98855}{Lehmann (1999)} for a precise
+#'   statement and conditions.
 #'
 #'   This movie considers examples where this limiting result holds and
 #'   illustrates graphically the closeness of the limiting approximation
 #'   provided by the relevant normal limit to the true finite-\eqn{n}
-#'   distribution.  Of course, when \code{distn = "normal"} this result is
-#'   exact.
+#'   distribution.
 #'
 #'   Samples of size \code{n} are repeatedly simulated from the distribution
 #'   chosen using \code{distn}.  These samples are summarized using a plot
-#'   that appears at the top of the movie screen.  For each sample the mean
-#'   of these \code{n} values is calculated, stored and added to another plot,
-#'   situated below the first plot.
+#'   that appears at the top of the movie screen.  For each sample the
+#'   100\eqn{p}\% sample quantile of these \code{n} values is calculated,
+#'   stored and added to another plot, situated below the first plot.
 #'   This plot is either a histogram or an empirical c.d.f., chosen using a
 #'   radio button.
 #'   A \code{\link[graphics]{rug}} is added to a histogram provided that it
 #'   contains no more than 1000 points.
 #'
-#'   The p.d.f. (for a continuous variable) or p.m.f. (for a discrete variable)
-#'   of the original variables is added to the top plot.
+#'   The p.d.f. of the original variables is added to the top plot.
 #'
 #'   Once it starts, four aspects of this movie are controlled by the user.
 #'   \itemize{
 #'     \item{}{There are buttons to increase (+) or decrease (-) the sample
-#'       size, that is, the number of values over which a mean is
+#'       size, that is, the number of values for which a sample quantile is
 #'       calculated.}
 #'     \item{}{Each time the button labelled "simulate another \code{n_add}
 #'       samples of size n" is clicked \code{n_add} new samples are simulated
-#'       and their sample mean are added to the bottom histogram.}
+#'       and their sample quantile are added to the bottom histogram.}
 #'     \item{}{There is a button to switch the bottom plot from displaying
-#'       a histogram of the simulated means and the limiting normal p.d.f. to
-#'       the empirical c.d.f. of the simulated data and the limiting normal
-#'       c.d.f.}
+#'       a histogram of the simulated sample quantiles and the limiting normal
+#'       p.d.f. to the empirical c.d.f. of the simulated data and the limiting
+#'       normal c.d.f.}
 #'     \item{}{There is a checkbox to add to the bottom plot the approximate
-#'       (large \code{n}) normal p.d.f./c.d.f. (with mean \eqn{\mu} and
-#'       standard deviation \eqn{\sigma / \sqrt{n}}), implied by the CLT.}
+#'       (large \eqn{n}) normal p.d.f./c.d.f. implied by the CLT for sample
+#'       quantiles: the mean is equal to \eqn{\xi(p)} and standard deviation is
+#'       equal to \eqn{\sqrt p \sqrt q  / n f(\xi(p))}, where \eqn{q = 1-p}}.
 #'   }
 #' @return Nothing is returned, only the animation is produced.
+#' @references Lehman, E. L. (1999) \emph{Elements of Large-Sample Theory},
+#'   Springer-Verlag, London. \url{http://dx.doi.org/10.1007/b98855}
 #' @seealso \code{\link{movies}}: a user-friendly menu panel.
 #' @seealso \code{\link{smovie}}: general information about smovie.
-#' @seealso \code{\link{cltq}}: Central Limit Theorem for sample quantiles.
+#' @seealso \code{\link{clt}}: Central Limit Theorem.
 #' @examples
 #' # Exponential data
-#' clt()
+#' cltq()
 #'
 #' # Uniform data
-#' clt(distn = "uniform")
-#'
-#' # Poisson data
-#' clt(distn = "poisson")
+#' cltq(distn = "t", params = list(df = 2))
 #' @export
-clt <- function(n = 20, distn, params = list(), panel_plot = TRUE, hscale = NA,
-                vscale = hscale, n_add = 1, delta_n = 1, arrow = TRUE,
-                leg_cex = 1.25, ...) {
+cltq <- function(n = 20, p = 0.5, distn, params = list(), type = 7,
+                 panel_plot = TRUE, hscale = NA, vscale = hscale, n_add = 1,
+                 delta_n = 1, arrow = TRUE, leg_cex = 1.25, ...) {
   if (!tcltk::is.tclObj(tcltk::tclRequire("BWidget"))) {
     message("Package BWidget was not found.")
     message("Please see the smovie README file for information.")
@@ -143,8 +142,11 @@ clt <- function(n = 20, distn, params = list(), panel_plot = TRUE, hscale = NA,
   vscale <- temp$vscale
   # To add another distribution
   # 1. misc.R: add code to set_fun_args(), set_top_range(), set_leg_pos()
-  # 2. add lines to rfun, dfun, qfun, pfun, distn_mean and distn_sd
-  # 3. cltmovie_plot(): add to the_distn
+  # 2. add lines to rfun, dfun, qfun, pfun, distn_den and distn_sd
+  # 3. cltqmovie_plot(): add to the_distn
+  if (p <= 0 || p >= 1) {
+    stop("p must be in (0, 1)")
+  }
   if (!is.wholenumber(n)) {
     stop("n must be an integer")
   }
@@ -183,11 +185,6 @@ clt <- function(n = 20, distn, params = list(), panel_plot = TRUE, hscale = NA,
            "chi-squared" = stats::rchisq,
            "f" = stats::rf,
            "weibull" = stats::rweibull,
-           "binomial" = stats::rbinom,
-           "geometric" = stats::rgeom,
-           "hypergeometric" = stats::rhyper,
-           "negative binomial" = stats::rnbinom,
-           "poisson" = stats::rpois,
            "gev" = revdbayes::rgev,
            NULL)
   if (is.null(rfun)) {
@@ -206,11 +203,6 @@ clt <- function(n = 20, distn, params = list(), panel_plot = TRUE, hscale = NA,
            "chi-squared" = stats::dchisq,
            "f" = stats::df,
            "weibull" = stats::dweibull,
-           "binomial" = stats::dbinom,
-           "geometric" = stats::dgeom,
-           "hypergeometric" = stats::dhyper,
-           "negative binomial" = stats::dnbinom,
-           "poisson" = stats::dpois,
            "gev" = revdbayes::dgev)
   qfun <-
     switch(distn,
@@ -225,11 +217,6 @@ clt <- function(n = 20, distn, params = list(), panel_plot = TRUE, hscale = NA,
            "chi-squared" = stats::qchisq,
            "f" = stats::qf,
            "weibull" = stats::qweibull,
-           "binomial" = stats::qbinom,
-           "geometric" = stats::qgeom,
-           "hypergeometric" = stats::qhyper,
-           "negative binomial" = stats::qnbinom,
-           "poisson" = stats::qpois,
            "gev" = revdbayes::qgev)
   pfun <-
     switch(distn,
@@ -244,36 +231,11 @@ clt <- function(n = 20, distn, params = list(), panel_plot = TRUE, hscale = NA,
            "chi-squared" = stats::pchisq,
            "f" = stats::pf,
            "weibull" = stats::pweibull,
-           "binomial" = stats::pbinom,
-           "geometric" = stats::pgeom,
-           "hypergeometric" = stats::phyper,
-           "negative binomial" = stats::pnbinom,
-           "poisson" = stats::ppois,
            "gev" = revdbayes::pgev)
   #
   discrete_distn <- FALSE
-  if (distn %in% c("binomial", "geometric", "hypergeometric",
-                   "negative binomial", "poisson")) {
-    discrete_distn <- TRUE
-  }
   # Set the arguments to the distributional functions
   fun_args <- set_fun_args(distn, dfun, fun_args, params)
-  # Check for finite variance
-  if (distn == "f") {
-    if (fun_args$df2 <= 4) {
-      stop("df2 must be greater than 4 for a finite variance")
-    }
-  }
-  if (distn == "t") {
-    if (fun_args$df <= 2) {
-      stop("df must be greater than 2 for a finite variance")
-    }
-  }
-  if (distn == "gp" || distn == "gev") {
-    if (fun_args$shape >= 1/2) {
-      stop("shape must be less than 1/2 for a finite variance")
-    }
-  }
   # Set sensible scales for the plots
   if (distn == "t") {
     if (fun_args$df < 2) {
@@ -310,12 +272,9 @@ clt <- function(n = 20, distn, params = list(), panel_plot = TRUE, hscale = NA,
   # Set a unique panel name to enable saving of objects to the correct panel
   now_time <- strsplit(substr(date(), 12, 19), ":")[[1]]
   now_time <- paste(now_time[1], now_time[2], now_time[3], sep = "")
-  my_panelname <- paste("clt_", now_time, sep = "")
+  my_panelname <- paste("cltq_", now_time, sep = "")
   old_n <- 0
   # Store the mean and standard deviation of the underlying distribution
-  if (distn == "hypergeometric") {
-    hp <- fun_args$m / (fun_args$m + fun_args$n)
-  }
   if (distn == "beta") {
     alpha <- fun_args$shape1
     beta <- fun_args$shape2
@@ -332,103 +291,44 @@ clt <- function(n = 20, distn, params = list(), panel_plot = TRUE, hscale = NA,
     nu <- fun_args$df
     ncp <- fun_args$ncp
   }
-  if (distn == "gev") {
-    mu <- fun_args$loc
-    sigma <- fun_args$scale
-    xi <- fun_args$shape
-    if (xi == 0) {
-      gev_mean <- mu - digamma(1) * sigma
-      gev_var <- sigma ^ 2 / 6
-    } else {
-      gev_mean <- mu + sigma * (gamma(1 - xi) - 1) / xi
-      gev_var <- sigma ^ 2 * (gamma(1 - 2 * xi) - gamma(1 - xi) ^ 2) / xi ^ 2
-    }
-  }
-  distn_mean <- switch(distn,
-                       "binomial" = fun_args$size * fun_args$prob,
-                       "geometric" = (1 - fun_args$prob) / fun_args$prob,
-                       "hypergeometric" = fun_args$k * hp,
-                       "negative binomial" = fun_args$size *
-                         (1 - fun_args$prob) / fun_args$prob,
-                       "poisson" = fun_args$lambda,
-                       "beta" = alpha / (alpha + beta),
-                       "chi-squared" = fun_args$df + fun_args$ncp,
-                       "f" = df2 * (df1 + lam) / (df1 * (df2 - 2)),
-                       "gamma" = fun_args$shape / fun_args$rate,
-                       "gp" = fun_args$loc + fun_args$scale /
-                         (1 - fun_args$shape),
-                       "gev" = gev_mean,
-                       "lognormal" = exp(fun_args$meanlog +
-                                           fun_args$sdlog ^ 2 / 2),
-                       "normal" = fun_args$mean,
-                       "t" = ncp * sqrt(nu / 2) * gamma((nu - 1) / 2) /
-                         gamma(nu / 2),
-                       "uniform" = (fun_args$min + fun_args$max) / 2,
-                       "weibull" = fun_args$scale *
-                         gamma(1 + 1 / fun_args$shape),
-                       "exponential" = 1 / fun_args$rate)
-  distn_sd <- switch(distn,
-                     "binomial" = sqrt(fun_args$size * fun_args$prob *
-                       (1 -fun_args$prob)),
-                     "geometric" = sqrt(1 - fun_args$prob) / fun_args$prob,
-                     "hypergeometric" = sqrt(fun_args$k * hp * (1 - hp) *
-                       (fun_args$m + fun_args$n - fun_args$k) /
-                         (fun_args$m + fun_args$n - 1)),
-                     "negative binomial" = sqrt(fun_args$size *
-                       (1 - fun_args$prob)) / fun_args$prob,
-                     "poisson" = sqrt(fun_args$lambda),
-                     "beta" = sqrt(alpha * beta / (alpha + beta) ^ 2 /
-                       (alpha + beta + 1)),
-                     "chi-squared" = sqrt(2 * (fun_args$df
-                                               + 2 * fun_args$ncp)),
-                     "f" = sqrt(num / den),
-                     "gamma" = sqrt(fun_args$shape / fun_args$rate ^ 2),
-                     "gp" = sqrt(fun_args$scale ^ 2 /
-                                   (1 - fun_args$shape) ^ 2 /
-                                   (1 - 2 * fun_args$shape)),
-                     "gev" = sqrt(gev_var),
-                     "lognormal" = sqrt(exp(2 * fun_args$meanlog +
-                                         fun_args$sdlog ^ 2) *
-                       (exp(fun_args$sdlog ^ 2) - 1)),
-                     "normal" = fun_args$sd,
-                     "t" = sqrt(nu * (1 + ncp ^ 2) / (nu - 2) + ncp ^ 2 * nu *
-                       (gamma((nu - 1) / 2) - gamma(nu / 2)) ^ 2 /2),
-                     "uniform" = sqrt((fun_args$max - fun_args$min) ^ 2 / 12),
-                     "weibull" = sqrt(fun_args$scale ^ 2 *
-                       gamma(1 + 2 / fun_args$shape) -
-                       gamma(1 + 1 / fun_args$shape) ^ 2),
-                     "exponential" = 1 / fun_args$rate)
+  # distn_den is the value of the density at the 100p% quantile
+  distn_den <- do.call(qfun, c(list(p = p), fun_args))
+  distn_sd <- sqrt(p * (1 - p)) /
+    do.call(dfun, c(list(x = distn_den), fun_args))
   # Create buttons for movie
   show_dens <- FALSE
   pdf_or_cdf <- "pdf"
   old_y <- NULL
   save_last_y <- NULL
-  sample_means <- NULL
-  cltpanel <- rpanel::rp.control("central limit theorem",
-                                 panelname = my_panelname, n = n, n_add = n_add,
-                                 dfun = dfun, qfun = qfun, rfun = rfun,
-                                 pfun = pfun, fun_args = fun_args,
-                                 distn = distn, top_range = top_range,
-                                 top_p_vec = top_p_vec,
-                                 bottom_p_vec = bottom_p_vec,
-                                 pdf_or_cdf = pdf_or_cdf, show_dens = show_dens,
-                                 top_leg_pos = top_leg_pos,
-                                 bottom_leg_pos = bottom_leg_pos,
-                                 xlab = xlab, arrow = arrow,
-                                 distn_mean = distn_mean, distn_sd = distn_sd,
-                                 discrete_distn = discrete_distn,
-                                 old_n = old_n, old_pdf_or_cdf = pdf_or_cdf,
-                                 old_show_dens = show_dens,
-                                 old_y = old_y, save_last_y = save_last_y,
-                                 sample_means = sample_means,
-                                 leg_cex = leg_cex)
+  sample_qs <- NULL
+  cltqpanel <- rpanel::rp.control("central limit theorem for sample quantiles",
+                                  panelname = my_panelname, n = n,
+                                  n_add = n_add,
+                                  dfun = dfun, qfun = qfun, rfun = rfun,
+                                  pfun = pfun, fun_args = fun_args,
+                                  distn = distn, top_range = top_range,
+                                  top_p_vec = top_p_vec,
+                                  bottom_p_vec = bottom_p_vec,
+                                  pdf_or_cdf = pdf_or_cdf,
+                                  show_dens = show_dens,
+                                  top_leg_pos = top_leg_pos,
+                                  bottom_leg_pos = bottom_leg_pos,
+                                  xlab = xlab, arrow = arrow,
+                                  distn_den = distn_den, distn_sd = distn_sd,
+                                  p = p,  type = type,
+                                  discrete_distn = discrete_distn,
+                                  old_n = old_n, old_pdf_or_cdf = pdf_or_cdf,
+                                  old_show_dens = show_dens,
+                                  old_y = old_y, save_last_y = save_last_y,
+                                  sample_qs = sample_qs,
+                                  leg_cex = leg_cex)
   #
   redraw_plot <- NULL
   panel_redraw <- function(panel) {
     rpanel::rp.tkrreplot(panel = panel, name = redraw_plot)
     # rp.tkrreplot() doesn't update the panel automatically, so do it manually
     # Get ...
-    panel$sample_means <- rpanel::rp.var.get(my_panelname, "sample_means")
+    panel$sample_qs <- rpanel::rp.var.get(my_panelname, "sample_qs")
     panel$old_n <- rpanel::rp.var.get(my_panelname, "old_n")
     panel$old_pdf_or_cdf <- rpanel::rp.var.get(my_panelname, "old_pdf_or_cdf")
     panel$old_show_dens <- rpanel::rp.var.get(my_panelname, "old_show_dens")
@@ -443,15 +343,15 @@ clt <- function(n = 20, distn, params = list(), panel_plot = TRUE, hscale = NA,
     panel_plot <- FALSE
   }
   if (panel_plot) {
-    rpanel::rp.tkrplot(panel = cltpanel, name = redraw_plot,
-                       plotfun = cltmovie_plot, pos = "right",
+    rpanel::rp.tkrplot(panel = cltqpanel, name = redraw_plot,
+                       plotfun = cltqmovie_plot, pos = "right",
                        hscale = hscale, vscale = vscale, background = "white")
     action <- panel_redraw
   } else {
-    action <- cltmovie_plot
+    action <- cltqmovie_plot
   }
   #
-  rpanel::rp.doublebutton(panel = cltpanel, variable = n, step = delta_n,
+  rpanel::rp.doublebutton(panel = cltqpanel, variable = n, step = delta_n,
                           title = "sample size, n",
                           action = action, initval = n,
                           range = c(2, NA), showvalue = TRUE, ...)
@@ -464,32 +364,33 @@ clt <- function(n = 20, distn, params = list(), panel_plot = TRUE, hscale = NA,
   # If the user hasn't set either repeatdelay or repeatinterval then set them
   # to the default values in rp.doublebutton (100 milliseconds)
   if (is.null(dlist$repeatdelay) & is.null(dlist$repeatinterval)) {
-      rpanel::rp.button(panel = cltpanel, action = action, title = my_title,
+      rpanel::rp.button(panel = cltqpanel, action = action, title = my_title,
                         repeatdelay = 100, repeatinterval = 100, ...)
   } else {
-    rpanel::rp.button(panel = cltpanel, action = action, title = my_title,
+    rpanel::rp.button(panel = cltqpanel, action = action, title = my_title,
                       ...)
   }
-  rpanel::rp.radiogroup(panel= cltpanel, pdf_or_cdf, c("pdf", "cdf"),
+  rpanel::rp.radiogroup(panel= cltqpanel, pdf_or_cdf, c("pdf", "cdf"),
                         title = "pdf or cdf in bottom plot", action = action)
-  rpanel::rp.checkbox(panel = cltpanel, show_dens,
+  rpanel::rp.checkbox(panel = cltqpanel, show_dens,
                       labels = "show normal pdf/cdf", action = action)
   if (!panel_plot) {
-    rpanel::rp.do(panel = cltpanel, action = action)
+    rpanel::rp.do(panel = cltqpanel, action = action)
   }
   return(invisible())
 }
 
-# Function to be called by clt().
+# Function to be called by cltq().
 
-cltmovie_plot <- function(panel) {
+cltqmovie_plot <- function(panel) {
   oldpar <- graphics::par(mfrow = c(2, 1), oma = c(0, 0, 0, 0),
                           mar = c(4, 4, 2, 2) + 0.1)
   on.exit(graphics::par(oldpar))
   # To please R CMD check
   n <- distn <- fun_args <- pdf_or_cdf <- show_dens <- n_add <- rfun <-
     discrete_distn <- top_range <- dfun <- xlab <- top_leg_pos <- arrow <-
-    distn_mean <- distn_sd <- bottom_p_vec <- bottom_leg_pos <- leg_cex <- NULL
+    distn_den <- distn_sd <- p <- type <- bottom_p_vec <- bottom_leg_pos <-
+    leg_cex <- NULL
   panel <- within(panel, {
     # Don't add the rug in the top plot if n is large
     if (n > 1000) {
@@ -498,29 +399,25 @@ cltmovie_plot <- function(panel) {
       show_rug <- TRUE
     }
     # Do the simulation (if required)
-    if (distn == "hypergeometric") {
-      sim_list <- c(list(nn = n), fun_args)
-    } else {
-      sim_list <- c(list(n = n), fun_args)
-    }
+    sim_list <- c(list(n = n), fun_args)
     if (old_pdf_or_cdf == pdf_or_cdf & old_show_dens == show_dens) {
       temp <- as.matrix(replicate(n_add, do.call(rfun, sim_list)))
-      mean_y <- apply(temp, 2, mean)
+      q_y <- apply(temp, 2, stats::quantile, probs = p, type = type)
       # Extract the last dataset and the last mean (for drawing the arrow)
       y <- temp[, n_add]
       old_y <- y
       rm(temp)
-      last_y <- mean_y[n_add]
+      last_y <- q_y[n_add]
       save_last_y <- last_y
     } else {
-      mean_y <- NULL
+      q_y <- NULL
       y <- old_y
       last_y <- save_last_y
     }
     if (n != old_n) {
-      sample_means <- mean_y
+      sample_qs <- q_y
     } else {
-      sample_means <- c(sample_means, mean_y)
+      sample_qs <- c(sample_qs, q_y)
     }
     #
     n_x_axis <- 501
@@ -562,13 +459,6 @@ cltmovie_plot <- function(panel) {
         "weibull" = paste("Weibull", "(", fun_args$shape, ",", fun_args$scale,
                           ")"),
         "t" = paste("Weibull", "(", fun_args$df, ",", fun_args$ncp, ")"),
-        "binomial" = paste(distn, "(", fun_args$size, ",", fun_args$prob, ")"),
-        "geometric" = paste(distn, "(", fun_args$prob, ")"),
-        "hypergeometric" = paste(distn, "(", fun_args$m, ",", fun_args$n,
-                      ",", fun_args$k, ")"),
-        "negative binomial" = paste(distn, "(", fun_args$size, ",",
-                                    fun_args$prob, ")"),
-        "poisson" = paste("Poisson", "(", fun_args$lambda, ")"),
         "gev" = paste("GEV", "(", fun_args$loc, ",", fun_args$scale,
                        ",", fun_args$shape, ")")
       )
@@ -614,12 +504,12 @@ cltmovie_plot <- function(panel) {
     #
     # Bottom plot --------
     #
-    my_xlab <- paste("sample mean of", n, "values")
-    # Set the mean and variance of the sample means parameters
-    normal_pars <- list(mean = distn_mean, sd = distn_sd / sqrt(n))
+    my_xlab <- paste0("sample ", 100 * p, "% quantile of ", n, " values")
+    # Set the mean and variance of the sample quantile parameters
+    normal_pars <- list(mean = distn_den, sd = distn_sd / sqrt(n))
     for_qnorm <- c(list(p = bottom_p_vec), normal_pars)
     normal_bottom_range <- do.call(stats::qnorm, for_qnorm)
-    bottom_range <- range(normal_bottom_range, sample_means)
+    bottom_range <- range(normal_bottom_range, sample_qs)
     # Set range for x-axis
     x <- seq(bottom_range[1], bottom_range[2], len = n_x_axis)
     # Calcuate the density over this range
@@ -641,8 +531,8 @@ cltmovie_plot <- function(panel) {
       ytop <- 1
     }
     # Histogram with rug
-    y <- sample_means
-    if (length(sample_means) > 1000) {
+    y <- sample_qs
+    if (length(sample_qs) > 1000) {
       show_bottom_rug <- FALSE
     } else {
       show_bottom_rug <- TRUE
@@ -673,7 +563,7 @@ cltmovie_plot <- function(panel) {
     }
     graphics::rug(last_y, line = 0.5, ticksize = 0.05, col = "red", lwd = 2)
     u_b <- my_xlim
-    my_leg_2 <- paste("N (", signif(distn_mean, 2), ",",
+    my_leg_2 <- paste("N (", signif(distn_den, 2), ",",
                       signif(distn_sd ^ 2, 2), "/ n )" )
     if (pdf_or_cdf == "pdf") {
       if (show_dens) {